extract keylistener and debug views from main
